use the vcs information from ReadBuildInfo in Go 1.18
